+++ driverlib/f28003x/driverlib/asysctl.h
@@ -342,7 +342,7 @@ static inline void ASysCtl_selectInternalTestNode(ASysCtl_TestSelect testSelect)
     HWREG(ANALOGSUBSYS_BASE + ASYSCTL_O_INTERNALTESTCTL) = 
         (HWREG(ANALOGSUBSYS_BASE + ASYSCTL_O_INTERNALTESTCTL) &
         ~(ASYSCTL_INTERNALTESTCTL_TESTSEL_M | ASYSCTL_INTERNALTESTCTL_KEY_M)) | 
-        (0xA5A50000UL | testSelect);
+        (0xA5A50000UL | (uint32_t)testSelect);
 
     EDIS;
 }
