+++ driverlib/f2837xd/driverlib/adc.h
@@ -6,7 +6,7 @@
 //
 //###########################################################################
 // 
-// C2000Ware v26.00.00.00
+// C2000Ware v26.01.00.00
 //
 // Copyright (C) 2024 Texas Instruments Incorporated - http://www.ti.com
 //
@@ -784,7 +784,7 @@ ADC_getInterruptStatus(uint32_t base, ADC_IntNumber adcIntNum)
     //
     // Get the specified ADC interrupt status.
     //
-    return((HWREGH(base + ADC_O_INTFLG) & (1U << (uint16_t)adcIntNum)) != 0U);
+    return((HWREGH(base + ADC_O_INTFLG) & ((uint16_t)1U << (uint16_t)adcIntNum)) != 0U);
 }
 
 //*****************************************************************************
@@ -850,7 +850,7 @@ ADC_getInterruptOverflowStatus(uint32_t base, ADC_IntNumber adcIntNum)
     //
     // Get the specified ADC interrupt status.
     //
-    return((HWREGH(base + ADC_O_INTOVF) & (1U << (uint16_t)adcIntNum)) != 0U);
+    return((HWREGH(base + ADC_O_INTOVF) & ((uint16_t)1U << (uint16_t)adcIntNum)) != 0U);
 }
 
 //*****************************************************************************
@@ -1666,7 +1666,7 @@ ADC_enableInterrupt(uint32_t base, ADC_IntNumber adcIntNum)
     //
     EALLOW;
 
-    HWREGH(intRegAddr) |= ADC_INTSEL1N2_INT1E << shiftVal;
+    HWREGH(intRegAddr) |= ((uint16_t)(ADC_INTSEL1N2_INT1E << shiftVal));
 
     EDIS;
 }
@@ -1813,7 +1813,7 @@ ADC_enableContinuousMode(uint32_t base, ADC_IntNumber adcIntNum)
     //
     EALLOW;
 
-    HWREGH(intRegAddr) |= ADC_INTSEL1N2_INT1CONT << shiftVal;
+    HWREGH(intRegAddr) |= ((uint16_t)(ADC_INTSEL1N2_INT1CONT << shiftVal));
 
     EDIS;
 }
