+++ driverlib/f28p65x/driverlib/adc.c
@@ -6,7 +6,7 @@
 //
 //###########################################################################
 // 
-// C2000Ware v26.00.00.00
+// C2000Ware v26.01.00.00
 //
 // Copyright (C) 2024 Texas Instruments Incorporated - http://www.ti.com
 //
@@ -242,7 +242,7 @@ ADC_setINLTrim(uint32_t base)
         EALLOW;
         for(i = 0U; i < 6U; i++)
         {
-            HWREG(base + ADC_O_INLTRIM1 + (i * 2U)) = (*inlTrimAddress++);
+            HWREG(base + ADC_O_INLTRIM1 + ((uint32_t)i * 2U)) = (*inlTrimAddress++);
         }
 
         //
@@ -420,7 +420,7 @@ ADC_configureRepeater(uint32_t base, uint16_t repInstance,
     ASSERT(repInstance <= 2U);
     ASSERT(config->repCount <= 127U);
 
-    regOffset = base + (repInstance * (ADC_REPxCTL_STEP));
+    regOffset = base + ((uint32_t)repInstance * (ADC_REPxCTL_STEP));
     EALLOW;
     //
     // Configure repeater mode, trigger and syncin source.
