+++ driverlib/f28p65x/driverlib/clb.c
@@ -6,7 +6,7 @@
 //
 //###########################################################################
 // 
-// C2000Ware v26.00.00.00
+// C2000Ware v26.01.00.00
 //
 // Copyright (C) 2024 Texas Instruments Incorporated - http://www.ti.com
 //
@@ -95,7 +95,7 @@ void CLB_clearFIFOs(uint32_t base)
 
     for(i = 0U; i < CLB_FIFO_SIZE; i++)
     {
-        HWREG(base + CLB_DATAEXCH + CLB_O_PULL(i)) = 0U;
+        HWREG(base + CLB_DATAEXCH + (uint32_t)CLB_O_PULL(i)) = 0U;
     }
 
     HWREG(base + CLB_LOGICCTL + CLB_O_BUF_PTR) = 0U;
