+++ driverlib/f28p65x/driverlib/dac.c
@@ -6,7 +6,7 @@
 //
 //###########################################################################
 // 
-// C2000Ware v26.00.00.00
+// C2000Ware v26.01.00.00
 //
 // Copyright (C) 2024 Texas Instruments Incorporated - http://www.ti.com
 //
@@ -72,7 +72,7 @@ DAC_tuneOffsetTrim(uint32_t base, float32_t referenceVoltage)
     // voltage other than 2.5v.
     //
     newOffsetTrim = ((float32_t)(2.5 / referenceVoltage) *
-                     (int16_t)oldOffsetTrim);
+                     (float32_t)(int16_t)oldOffsetTrim);
 
     //
     // Check if the new offset trim value is valid
@@ -85,7 +85,7 @@ DAC_tuneOffsetTrim(uint32_t base, float32_t referenceVoltage)
     EALLOW;
     HWREGH(base + DAC_O_TRIM) = (HWREGH(base + DAC_O_TRIM) &
                                  ~DAC_TRIM_OFFSET_TRIM_M) |
-                                 (int16_t)newOffsetTrim;
+                                 (uint16_t)(int16_t)newOffsetTrim;
 
     EDIS;
